updated cout statements
Former-commit-id: 25fde0779c92ea57d104e3b6eeb8d18d076555db

+++ branches/trunk/algorithms/BasicTracking.C
@@ -202,7 +202,7 @@ StatusCode BasicTracking::run(Clipboard* clipboard){
   }
 
   nTracksTotal+=tracks->size();
-  cout<<", produced "<<nTracksTotal<<" tracks";
+  cout<<", produced "<<(int)nTracksTotal<<" tracks";
 
   // Clean up tree objects
 //  for(int det = 0; det<parameters->nDetectors; det++){
